Containerized app (#4)
Delete, Write functionality combined into one container (python flask app).
Read APIs folder is prepared to be a go backend (another cotainer). Part 1 code for package rater is also placed as another container.

Actions for linting and testing is included in yml. Black package will make soft changes on linting that needs to occur for python code. Pylint will rate (out of 10) the state of the code and will return failure if score is < 8.

In CI, each container is built successfully but the dockerfile is stripped from running the app because it will make the CI run indefinitely (server is running in the VM terminal).

Missing creation of integration test, handling of database, src code and unit tests for backend.